Fixed the <#root#>OFB</#root#> bug.
Basically, to recognize enumerations in the js, we rely on a
"hard-coded" Enum clafer. Padding and Mode did not extend this Enum
clafer so I fixed that. Also, since both of these do not have a
reference type (e.g., integer like security and performance), they need
a special case to handle them