Fixes #31.
Removed all files related to PWDEncryption task and removed it from the
tasks file. Added a new question to Symmetric-Encryption task about
whether the key should be derived from a password. If answered yes, a
kda is added to the required algorithms in the Clafer model and the
KeyDeriv class is being generated.

Along the way, added a createClafer() method in the ClaferModelUtils
class. It is presumably quite fragile and should be used with caution in
other circumstances.